% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_instances.R
\name{add_noise}
\alias{add_noise}
\title{Add noise...}
\usage{
add_noise(F, depth, overdispersion)
}
\arguments{
\item{F}{F matrix}

\item{depth}{depth}

\item{overdispersion}{overdispersion}
}
\value{
Matrix containing...
}
\description{
\code{add_noise} is ...
}
\details{
Aqui va la descripcion

Aqui sigue la descripcion
}
