% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{B_mats}
\alias{B_mats}
\title{A set of 10 trios of B matrices for experimenting with the methods of \code{GeRnika}}
\format{
A list of lists composed by 10 trios of B matrices.
\describe{
  \item{Trio 1}{B_real, B_alg1 and B_alg2 (matrices composed by 5 clones)}
  \item{Trio 2}{B_real, B_alg1 and B_alg2 (matrices composed by 5 clones)}
  \item{Trio 3}{B_real, B_alg1 and B_alg2 (matrices composed by 5 clones)}
  \item{Trio 4}{B_real, B_alg1 and B_alg2 (matrices composed by 5 clones)}
  \item{Trio 5}{B_real, B_alg1 and B_alg2 (matrices composed by 5 clones)}
  \item{Trio 6}{B_real, B_alg1 and B_alg2 (matrices composed by 10 clones)}
  \item{Trio 7}{B_real, B_alg1 and B_alg2 (matrices composed by 10 clones)}
  \item{Trio 8}{B_real, B_alg1 and B_alg2 (matrices composed by 10 clones)}
  \item{Trio 9}{B_real, B_alg1 and B_alg2 (matrices composed by 10 clones)}
  \item{Trio 10}{B_real, B_alg1 and B_alg2 (matrices composed by 10 clones)}
}
}
\source{
Local source; as a result of the Grasp and the ILS methods used for solving the Clonal Deconvolution and Evolution Problem (CDEP).
}
\usage{
B_mats
}
\description{
A list of lists composed by 10 trios of B matrices; a real B matrix, a B matrix got by using the one algorithm (alg1) method and another one as a result of another algorithm (alg2). These matrices can be used as examples for the methods of \code{GeRnika}.
}
\keyword{datasets}
