% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_instances.R
\name{calc_clone_proportions}
\alias{calc_clone_proportions}
\title{Calculate clone proportions...}
\usage{
calc_clone_proportions(B, selection)
}
\arguments{
\item{B}{B matrix}

\item{selection}{selection}
}
\value{
Tibble containing...
}
\description{
\code{calc_clone_proportions} is ...
}
\details{
Aqui va la descripcion

Aqui sigue la descripcion
}
