% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_instances.R
\name{create_B}
\alias{create_B}
\title{Create topology and B matrix}
\usage{
create_B(n, k)
}
\arguments{
\item{n}{number of clones.}

\item{k}{topology parameter}
}
\value{
Matrix containing...
}
\description{
\code{create_B} is ...
}
\details{
Aqui va la descripcion

Aqui sigue la descripcion
}
