% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_instances.R
\name{create_instance}
\alias{create_instance}
\title{Simulate tumor data}
\usage{
create_instance(
  n,
  m,
  k,
  selection,
  noisy = TRUE,
  depth = 30,
  seed = Sys.time()
)
}
\arguments{
\item{n}{The number of clones.}

\item{m}{The number of samples.}

\item{k}{Continuous number how branchy the created topology is.}

\item{selection}{Character that specifies the clone selection. Possible values: \code{"positive"} and \code{"neutral"}.}

\item{noisy}{An optional logical that specifies whether noise is added to values in \code{F} or not. FALSE by default.}

\item{depth}{An optional argument representing the read sequencing depth (for noisy cases). 30 by default.}

\item{seed}{An optional argument for controlling the randomness of the generated instances. Sys.time() by default.}
}
\value{
The instance of a tumor sample, composed by \code{F}, \code{F_true}, \code{B} and \code{U} .
}
\description{
Simulates a tumor instance, composed by \code{F}, \code{F_true}, \code{B} and \code{U}.
}
\examples{
# Create an instance composed by 10 clones,
# 4 samples, k = 1, "neutral" selection and
# with added noise and depth = 500
I1 <- create_instance(
  n = 10,
  m = 4,
  k = 1,
  selection = "neutral",
  depth = 500)
  

# Create an instance composed by 50 clones,
# 10 samples, k = 5, "positive" selection with 
# added noise and depth = 500
I2 <- create_instance(
  n = 50,
  m = 10,
  k = 5,
  selection = "positive", 
  noisy = TRUE,
  depth = 500)
  
  
# Create an instance composed by 100 clones,
# 25 samples, k = 0, "positive" selection without 
# added noise
I3 <- create_instance(
  n = 100,
  m = 25,
  k = 0,
  selection = "positive", 
  noisy = FALSE)
}
