% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree_class.R
\name{create_phylotree}
\alias{create_phylotree}
\title{Create a \code{Phylotree} object.}
\usage{
create_phylotree(B, clones, genes, parents, tree, labels = NA)
}
\arguments{
\item{B}{The Matrix that represents the clones in the phylogenetic tree.}

\item{clones}{numeric vector representing the clones in the phylogenetic tree.}

\item{genes}{numeric vector representing the genes in the phylogenetic tree.}

\item{parents}{numeric vector representing the parents the clones in the phylogenetic tree.}

\item{tree}{\code{data.tree} object containing the tree structure of the phylogenetic tree.}

\item{labels}{Optional argument that refers to the list containing the tags of the genes of the phylogenetic tree. \code{NA} by default.}
}
\value{
A \code{Phylotree} class object.
}
\description{
The general constructor of the \code{Phylotree} S4 class.
}
\examples{
# Create a B matrix instance
# composed by 10 subpopulations of
# clones
B <- create_instance(
       n = 10, 
       m = 4, 
       k = 1, 
       selection = "neutral")$B


# Create a new 'Phylotree' object
# on the basis of the B matrix
phylotree1 <- B_to_phylotree(B = B)


# Create a new 'Phylotree' object
# with the general constructor of
# the class
phylotree2 <- create_phylotree(
                B = B, 
                clones = tree@clones, 
                genes = tree@genes, 
                parents = tree@parents, 
                tree = tree@tree)


# Generate the tags for the genes of
# the phyogenetic tree
tags <- LETTERS[1:nrow(B)]

 
# Create a new 'Phylotree' object
# with the general constructor of
# the class using tags
phylotree_tags <- create_phylotree(
                    B = B, 
                    clones = tree@clones, 
                    genes = tree@genes, 
                    parents = tree@parents, 
                    tree = tree@tree, 
                    labels = tags)
}
