% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_instances.R
\name{create_U}
\alias{create_U}
\title{Create U matrix...}
\usage{
create_U(B, clone_proportions, density_coords, m, x)
}
\arguments{
\item{B}{B matrix}

\item{clone_proportions}{clone_proportions}

\item{density_coords}{density_coords}

\item{m}{number of samples}

\item{x}{x}
}
\value{
Matrix containing...
}
\description{
\code{create_U} is ...
}
\details{
Aqui va la descripcion

Aqui sigue la descripcion
}
