% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_instances.R
\name{.distribute_freqs}
\alias{.distribute_freqs}
\title{Calculate local frequencies in multifurcations}
\usage{
.distribute_freqs(B, clone_idx, clone_proportions, selection)
}
\arguments{
\item{B}{B matrix}

\item{clone_idx}{clone_idx}

\item{clone_proportions}{clone_proportions}

\item{selection}{selection}
}
\value{
Vector containing...
}
\description{
\code{.distribute_freqs} is ...
}
\details{
Aqui va la descripcion

Aqui sigue la descripcion
}
