% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree_class.R
\name{get_clones}
\alias{get_clones}
\title{Retrieve the clone indices for a set of gene indices}
\usage{
get_clones(genes)
}
\arguments{
\item{genes}{A vector of gene indices.}
}
\value{
A vector of clone indices.
}
\description{
Retrieve the clone indices for a set of gene indices
}
