% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree_class.R
\name{get_genes}
\alias{get_genes}
\title{Retrieve the gene indices for the clones in a phylogenetic tree}
\usage{
get_genes(B)
}
\arguments{
\item{B}{A square matrix that represents the phylogenetic tree.}
}
\value{
A vector of gene indices.
}
\description{
Retrieve the gene indices for the clones in a phylogenetic tree
}
