% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree_class.R
\name{get_parents}
\alias{get_parents}
\title{Get parent nodes in a phylogenetic tree.}
\usage{
get_parents(phylotree)
}
\arguments{
\item{phylotree}{An object of the \code{Phylotree} class representing the phylogenetic tree.}
}
\value{
A vector of parent nodes.
}
\description{
This function retrieves the parent nodes for each node in a given phylogenetic tree.
}
