% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hyperparameters}
\alias{hyperparameters}
\title{hyperparameters for the methods of \code{GeRnika}}
\format{
A data.frame containing different static values.
\describe{
  \item{Overdispersion}{value = 0.5}
  \item{Depth_sequencing}{value = 30.0}
}
}
\source{
local source; inspired on the optimal parameters for the methods of \code{GeRnika}.
}
\usage{
hyperparameters
}
\description{
A data.frame containing the static values for the parameters used in the methods of \code{GeRnika}
}
\keyword{datasets}
