% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree_class.R
\docType{class}
\name{Node-class}
\alias{Node-class}
\title{S4 class to represent a node in a phylogenetic tree.}
\description{
S4 class to represent a node in a phylogenetic tree.
}
\section{Slots}{

\describe{
\item{\code{Node}}{A node object}
}}

