% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{Palettes for the methods of \code{GeRnika}}
\format{
A data.frame containing 3 palettes.
\describe{
  \item{Lancet}{#0099B444, #AD002A77, #42B540FF}
  \item{NEJM}{#FFDC9177, #7876B188, #EE4C97FF}
  \item{Simpsons}{#FED43966, #FD744688, #197EC0FF}
}
}
\source{
Lancet, NEJM and The Simpsons palettes; inspired by the plots in Lancet journals, the plots in the New England Journal of Medicine and the colors used in the TV show The Simpsons, respectively.
}
\usage{
palettes
}
\description{
A data.frame containing 3 default palettes for the parameters used in the methods of \code{GeRnika}.
}
\keyword{datasets}
