% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree_class.R
\docType{class}
\name{Phylotree-class}
\alias{Phylotree-class}
\title{S4 class to represent phylogenetic trees.}
\description{
S4 class to represent phylogenetic trees.
}
\section{Slots}{

\describe{
\item{\code{B}}{A data.frame containing the square matrix that represents the ancestral relations among the clones of the phylogenetic tree.}

\item{\code{clones}}{A vector representing the equivalence table of the clones in the phylogenetic tree.}

\item{\code{genes}}{A vector representing the equivalence table of the genes in the phylogenetic tree.}

\item{\code{parents}}{A vector representing the parents of the clones in the phylogenetic tree.}

\item{\code{tree}}{A \code{Node} class object representing the phylogenetic tree.}

\item{\code{labels}}{A vector representing the tags of the genes in the phylogenetic tree.}
}}

