% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree_class.R
\docType{class}
\name{Phylotree_class}
\alias{Phylotree_class}
\title{Phylotree_class
S4 class to represent phylogenetic trees.}
\description{
Phylotree_class
S4 class to represent phylogenetic trees.
}
\section{Slots}{

\describe{
\item{\code{B}}{the data.frame containing the square matrix that represents the clones of the phylogenetic tree.}

\item{\code{clones}}{a vector representing the equivalence table of the clones in the phylogenetic tree.}

\item{\code{genes}}{a vector representing the equivalence table of the genes in the phylogenetic tree.}

\item{\code{parents}}{a vector representing the parents of the clones in the phylogenetic tree.}

\item{\code{tree}}{a \code{Node} class object representing the phylogenetic tree.}

\item{\code{labels}}{a vector representing the tags of the genes in the phylogenetic tree.}
}}

