% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree_class.R
\name{phylotree_to_tree}
\alias{phylotree_to_tree}
\title{Extract tree from \code{Phylotree} object.}
\usage{
phylotree_to_tree(phylotree)
}
\arguments{
\item{phylotree}{An object of the \code{Phylotree} class.}
}
\value{
The tree structure of the input \code{Phylotree} object.
}
\description{
This function extracts the tree structure from a given Phylotree object.
}
