% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_instances.R
\name{place_clones_space}
\alias{place_clones_space}
\title{Place clones in space...}
\usage{
place_clones_space(B)
}
\arguments{
\item{B}{B matrix}
}
\value{
List containing...
}
\description{
\code{place_clones_space} is ...
}
\details{
Aqui va la descripcion

Aqui sigue la descripcion
}
