% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree_class.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,Phylotree-method}
\title{Plot a Phylotree object.}
\usage{
plot(object, labels = FALSE)

\S4method{plot}{Phylotree}(object, labels = FALSE)
}
\arguments{
\item{object}{A \code{Phylotree} object.}

\item{labels}{A label vector.}
}
\description{
Plot a Phylotree object.
}
