% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree_class.R
\name{plot_p}
\alias{plot_p}
\title{Plot a phylogenetic tree with proportional node sizes and colors.}
\usage{
plot_p(phylotree, proportions)
}
\arguments{
\item{phylotree}{An object of the \code{Phylotree} class representing the phylogenetic tree to be plotted.}

\item{proportions}{A numeric vector representing the proportions of each clone in the phylogenetic tree. The length of this vector should be equal to the number of clones in the tree.}
}
\value{
A graph representing the phylogenetic tree, with node sizes and colors reflecting clone proportions.
}
\description{
This function plots a phylogenetic tree with nodes sized and colored according to the proportions of each clone.
}
