% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree_class.R
\name{plot_phylotree}
\alias{plot_phylotree}
\title{Plot a phylogenetic tree}
\usage{
plot_phylotree(phylotree, labels = FALSE)
}
\arguments{
\item{phylotree}{An object of the \code{Phylotree} class representing the phylogenetic tree to be plotted.}

\item{labels}{A logical value indicating whether to label the nodes with the labels stored in the \code{Phylotree} object. Default is FALSE.}
}
\value{
A plot of the phylogenetic tree.
}
\description{
This function generates a plot of a phylogenetic tree. If the `labels` parameter is set to TRUE, the nodes of the tree will be labeled with the labels stored in the \code{Phylotree} object.
}
