% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree_class.R
\name{plot_proportions}
\alias{plot_proportions}
\title{Plot a phylogenetic tree according to its clones' proportions}
\usage{
plot_proportions(phylotree, proportions, labels = FALSE)
}
\arguments{
\item{phylotree}{A \code{Phylotree} class object}

\item{proportions}{A vector containing the proportions of each clone in the phylogenetic tree. If a matrix is given, this method will plot more than one phylogenetic tree according to the proportions that each row contains.}

\item{labels}{A boolean, if \code{TRUE} the rendered graph will be plotted with the tags of the genes in the phylogenetic trees instead of their gene index. \code{FALSE} by default.}
}
\description{
Plots a phylogenetic tree according to its clones' proportions. It is possible to plot several phylogenetic trees.
}
\examples{
# Create an instance
# composed by 5 subpopulations of clones
# and 4 samples
instance <- create_instance(
       n = 5, 
       m = 4, 
       k = 1, 
       selection = "neutral")

# Create a new 'Phylotree' object
# on the basis of the B matrix
phylotree <- B_to_phylotree(B = B)

# Generate the tags for the genes of
# the phyogenetic tree
tags <- LETTERS[1:nrow(B)]

# Plot the phylogenetic tree taking
# into account the proportions of the
# previously generated instance
plot_proportions(phylotree, instance$U, labels=TRUE)
}
